# Conflicts:
#   src/main/java/teamproject/pocoapoco/service/CrewReviewService.java
#   src/main/resources/templates/alarms/alarm.html

+++ src/main/java/teamproject/pocoapoco/service/CrewService.java
@@ -33,6 +33,7 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 @Slf4j
 public class CrewService {
 
@@ -290,7 +291,11 @@ public class CrewService {
 
             Participation befParticipation = participationOptional.get();
 
-            participationRepository.delete(befParticipation);
+            if(befParticipation.getUser().getId().equals(actingUser.getId())){
+                throw new AppException(ErrorCode.NOT_AUTHORIZED, "방장은 강퇴할 수 없습니다.");
+            } else{
+                participationRepository.delete(befParticipation);
+            }
 
 
         }else{
